requirements.txt
blinker==1.6.3
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Flask==3.0.0
Flask-Login==0.6.2
Flask-OpenID==1.3.0
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
minio==7.1.17
pbr==5.11.1
python-dotenv==1.0.0
python3-openid==3.2.0
six==1.16.0
SQLAlchemy==2.0.22
sqlalchemy-migrate==0.13.0
sqlparse==0.4.4
Tempita==0.5.2
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
#